fix: add existence checks before modifying configuration files and streamline script execution

+++ .bin/mv/jar.sh
@@ -1,4 +1,4 @@
-if [ -f "./Floodgate-Spigot.jar" ] || [ -f "./Geyser-Spigot.jar" ] || [ -f "./Via-Version.jar" ]; then
+if [ ! -f "./plugins/Floodgate-Spigot.jar" ] && [ ! -f "./plugins/Geyser-Spigot.jar" ] && [ ! -f "./plugins/Via-Version.jar" ]; then
     mv ./Floodgate-Spigot.jar plugins/
     mv ./Geyser-Spigot.jar plugins/
     mv ./Via-Version.jar plugins/
@@ -7,4 +7,5 @@ if [ -f "./Floodgate-Spigot.jar" ] || [ -f "./Geyser-Spigot.jar" ] || [ -f "./Vi
     make reboot
     sleep 10
     make stop
+    sleep 10
 fi
